refactor(admin): AdminEnseignants sous 300 lignes — éclatement complet (#G1 ≤300)
AdminEnseignants.vue : 655 -> 192. Composable useAdminEnseignants (état + stats +
chargement KLASSCI + cache/refresh) ; sous-composants EnseignantsStatsCards et
EnseignantCard (carte de la grille) avec leur CSS. La vue orchestre header + états +
grille + EnseignantDetailModal (déjà extraite). CSS distribué ; overrides dark des
badges de la modale conservés dans la vue (parité, escapés via :global). CSS mort
tags-* retiré.

Tests : useAdminEnseignants (mock services) + EnseignantsStatsCards + EnseignantCard.
8/8 verts. vite build vert. Tous les fichiers ≤300.
